+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog - neom_settings
 
+## [1.1.0] - 2026-07-09
+- Update settings and privacy settings widgets.
+
+
 ## Unreleased - System updates
 - Actualizaciones de estabilidad y compatibilidad.